# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
...
  /admin/applications:
    get:
      operationId: getAdminApplications
      summary: List all loan applications
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationList"
  /admin/applications/{id}:
    get:
      operationId: getAdminApplication
      summary: Get a single application with documents and loan
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationDetail"
    patch:
      operationId: updateAdminApplication
      summary: Update application status / notes / approve
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # ── ADMIN: LOANS ──────────────────────────────────────
...
  /admin/documents:
    get:
      operationId: getAdminDocuments
      summary: List all KYC documents
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: applicationId
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /admin/documents/{id}:
    patch:
      operationId: updateAdminDocument
      summary: Approve or reject a KYC document
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  # ── ADMIN: NOTIFICATIONS ─────────────────────────────
...
components:
  schemas:
...
    Application:
      type: object
      required: [id, fullName, nationalId, phone, email, address, employmentType, employer, monthlyIncome, requestedAmount, repaymentMonths, payoutMethod, status, referenceNumber, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        nationalId:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        employmentType:
          type: string
        employer:
          type: string
        monthlyIncome:
          type: number
        requestedAmount:
          type: number
        repaymentMonths:
          type: integer
        payoutMethod:
          type: string
        status:
          type: string
        referenceNumber:
          type: string
        adminNotes:
          type: ["string", "null"]
        approvedAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        approvedAmount:
          type: ["number", "null"]
        sendSms:
          type: ["boolean", "null"]
    ApplicationDetail:
      allOf:
        - $ref: "#/components/schemas/Application"
        - type: object
          properties:
            documents:
              type: array
              items:
                $ref: "#/components/schemas/Document"
            loan:
              type: ["object", "null"]
              properties:
                id:
                  type: integer
                status:
                  type: string
                outstandingBalance:
                  type: number
    ApplicationList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Document:
      type: object
      required: [id, applicationId, docType, objectKey, fileName, status, uploadedAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        docType:
          type: string
        objectKey:
          type: string
        fileName:
          type: string
        fileSize:
          type: ["integer", "null"]
        status:
          type: string
        adminNotes:
          type: ["string", "null"]
        uploadedAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
    DocumentUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]

    # ── REPAYMENT CALCULATOR ─────────────────────────────
...
    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        fullName:
          type: ["string", "null"]

    # ── DASHBOARD ────────────────────────────────────────
...