# lib/api-zod/src/generated/types/applicationDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { Application } from './application';
import type { ApplicationDetailLoan } from './applicationDetailLoan';
import type { Document } from './document';

export type ApplicationDetail = Application & {
  documents?: Document[];
  /** @nullable */
  loan?: ApplicationDetailLoan;
};