# lib/api-zod/src/generated/types/applicationList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { Application } from './application';

export interface ApplicationList {
  items: Application[];
  total: number;
  page: number;
  limit: number;
}